Move constants from actions file
