+++ app/containers/Dashboard/constants.js
@@ -1,0 +1,5 @@
+// constants
+export const ADVANCED = 'advanced';
+export const OVERVIEW = 'overview';
+export const POWERUP = 'powerUp';
+export const POWERDOWN = 'powerDown';
